feat(Settings): start on new settings ui
