+++ src/components/Settings.tsx
@@ -4,6 +4,7 @@ import { DEFAULT_DATE_FORMAT, Settings as ISettings, Unit } from "types/Settings
 import { classes } from "lib/classes";
 import { Positions, Theme } from "lib/constants";
 import { FormField } from "./FormField";
+import { PositionButtons } from "./PositionButtons";
 
 interface Props {
   open: boolean;
@@ -292,37 +293,3 @@ export const Settings = ({ open, settings, onClose, onSettingsChange }: Props) =
     </ReactModal>
   );
 };
-
-interface PositionProps {
-  disabled: number;
-  onClick: (n: number) => void;
-  isActive: (n: number) => string;
-}
-
-function PositionButtons({ disabled, onClick, isActive }: PositionProps) {
-  return (
-    <div style={{ gap: "0.5em", display: "flex" }}>
-      <button
-        disabled={Positions.TOP_RIGHT === disabled}
-        onClick={() => onClick(Positions.TOP_RIGHT)}
-        className={classes("positionBtn", isActive(Positions.TOP_RIGHT))}
-      >
-        Top right
-      </button>
-      <button
-        disabled={Positions.BOTTOM_LEFT === disabled}
-        onClick={() => onClick(Positions.BOTTOM_LEFT)}
-        className={classes("positionBtn", isActive(Positions.BOTTOM_LEFT))}
-      >
-        Bottom Left
-      </button>
-      <button
-        disabled={Positions.BOTTOM_RIGHT === disabled}
-        onClick={() => onClick(Positions.BOTTOM_RIGHT)}
-        className={classes("positionBtn", isActive(Positions.BOTTOM_RIGHT))}
-      >
-        Bottom right
-      </button>
-    </div>
-  );
-}
