+++ src/components/Settings/Sidebar.tsx
@@ -1,3 +1,4 @@
+import { classes } from "lib/classes";
 import { Tabs } from "types/Settings";
 
 interface Props {
@@ -6,5 +7,19 @@ interface Props {
 }
 
 export const SettingsSidebar = ({ activeTab, setActiveTab }: Props) => {
-  return <div>Hello world!</div>;
+  const isActive = (n: number) => (activeTab === n ? "active" : "");
+
+  return (
+    <aside className="settingsSidebar">
+      <li className={classes("sidebarItem", isActive(Tabs.GENERAL))}>
+        <button onClick={() => setActiveTab(Tabs.GENERAL)}>General Setting</button>
+      </li>
+      <li className={classes("sidebarItem", isActive(Tabs.SEARCH))}>
+        <button onClick={() => setActiveTab(Tabs.SEARCH)}>Search Setting</button>
+      </li>
+      <li className={classes("sidebarItem", isActive(Tabs.WEATHER))}>
+        <button onClick={() => setActiveTab(Tabs.WEATHER)}>Weather Setting</button>
+      </li>
+    </aside>
+  );
 };
