+++ src/components/Settings/Tabs/General.tsx
@@ -1,3 +1,18 @@
+import { FormField } from "components/FormField";
+import { PositionButtons } from "components/PositionButtons";
+import { Positions } from "lib/constants";
+
 export const GeneralTab = () => {
-  return <h1>hello world from general tab</h1>;
+  function onGreetingClick() {}
+  const isGreetingActive = Positions.TOP_RIGHT;
+
+  return (
+    <div className="tab">
+      <form>
+        <FormField fieldId="greeting-position" label="Greeting Position">
+          <PositionButtons disabled={3} onClick={onGreetingClick} isActive={isGreetingActive} />
+        </FormField>
+      </form>
+    </div>
+  );
 };
